Move everything into more appropriate namespaces
