+++ common/src/main/scala/uk/ac/wellcome/platform/archive/common/verify/s3/S3ObjectVerifier.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.archive.common.storage.services
+package uk.ac.wellcome.platform.archive.common.verify.s3
 
 import java.net.URI
