+++ common/src/test/scala/uk/ac/wellcome/platform/archive/common/fixtures/VerifyFixtures.scala
@@ -1,7 +1,8 @@
 package uk.ac.wellcome.platform.archive.common.fixtures
 
-import uk.ac.wellcome.platform.archive.common.storage.services.{S3ObjectVerifier, S3Resolvable}
+import uk.ac.wellcome.platform.archive.common.storage.services.S3Resolvable
 import uk.ac.wellcome.platform.archive.common.verify._
+import uk.ac.wellcome.platform.archive.common.verify.s3.S3ObjectVerifier
 import uk.ac.wellcome.storage.ObjectLocation
 import uk.ac.wellcome.storage.fixtures.S3Fixtures
